Create description.txt
The index.html file contains the HTML markup for the image scroll setup, while the inline CSS styles and JavaScript code are included within the same file.

The images are displayed within a container element with an id of image-track. This container has a display property of flex, which allows the images to be displayed side-by-side in a row.

The JavaScript code listens for mouse events on the window object and updates the transform property of the image-track element to shift the position of the images based on the mouse movement.